Add Thor task to download, unpack, and set up eXist
This commit unpacks the eXist .war file into a new directory. Because
the webapp is exploded and the filepaths slightly differ, the new
task checks for older path references and updates them. It also
changes eXist's default ports to reflect Jetty's.

The new task is intended to configure eXist, removing the need for
cerberus_core's Exist Generator.

To run the task, use "thor exist_jetty:init".