Add CoreFilesController#nid_update
This allows us to track file revisions that are made by a TEI user to
one of their objects.  Also adds a convenience method for taking a form
upload object, moving its file to a location that won't get unlinked by
the controller when the request ends, and translating params[:file] into
a simple absolute filepath.