Edit eXist's default security permissions
When Jetty is running, the task will try to change eXist's default
admin password (an empty string) to "dsgT@pas". It will also
change the default database permissions from "rwxrwxrwx" to
"rwxrwxr--". This allows eXist to work fully out-of-the-box, since
eXist does not allow modifications while the admin password is
an empty string.