Use built in querying for nid methods
Previously, this module defined a find_by_nid method that used
ActiveFedora::SolrService::Query.  Since the .where method does the
trick and is less verbose and returns an object already cast to its
model (instead of a SolrDocument) we should use this instead.  Also adds
an exists_by_nid? method that checks if any object in the repository is
already using the given nid.