Add upserter service, implement for Communities
Given how drupal tapas works, it winds up being easier to implement a
rails endpoint called "upsert" that knows how to conditionally
create/update than to implement both as separate actions and case
between them in the drupal project.  This commit implements the job for
dispatching to a particular upserter based on the controller being used,
an elementary upserter concern, and the actual implementation of upsert
to be used by Communities.