Update validation logic
Object validations are now only run against the upsert action.  The only
other planned endpoints are for object deletion, which only require that
the user successfully authenticate when sending the DELETE request.
Validators and specs are also updated to reflect changes to the post
params required on inserting a new object into the repository.