Significant improvements to API authentication
Rather than require that users pass in a param[:email] and param[:token]
with their request, we now check for a token in the HTTP Authorization
header.  Because tokens are randomly generated 20 character sequences
and therefore somewhat stronger than the usual user password, we also
switch from BCrypt encrypted tokens to just a SHA512 hash of those
tokens, which simplifies + speeds up verification.  For our purposes I'm
mostly sure that this is secure enough.