Add ensure_unique_did
Adds a before_save callback that verifies that a Drupal ID isn't being
reused.  Because the obvious way of adding this callback (an included do
block in the did concern) simply doesn't appear to work I went ahead and
did the quick/dirty fix of explicitly adding it to each model.