Implement eXistdb MODS record retrieval action
Implements an action object that handles sending a TEI document to exist
and returning the MODS metadata extracted from it.  Also moves the very
small amount of shared functionality that exist_service implemented into
an exist_actions concern.  Updates configuration for dev/test
environments to declare an exist specific username/password combination
and to point at the right port, which is generally valuable.