Updates #all_ography_tei_files to expect CoreFile
Previously, this expected to be called on a Community (Project) object.
However, since we want to attach ographies to one or more Collections
instead (i.e. normal TEI content handling) it makes more sense to expect
this to be called on a CoreFile that can then be queried for its
containing collections.