Pass options hash to #add_unique_file
It is ocassionaly the case that we want to pass a blob and a filename
and occasionally the case that we want to pass a path to a file from
which a content blob/filename can be inferred.  Pass an options hash to
add_unique_file that lets us pick between the two.