Update validator file_types logic
File_types are no longer required on create, since instead of passing
'tei_content' we just assume that any tei doc that has no ography types
specified is only a primary document.  In addition, validation no longer
assumes that an object should ever be attached to a project directly.
We also handle the case where we are passed a single string for an
ography by casting it to an array.