Reimplement TransformHTMLURLs
It turns out to be the case that we need to process the XML that is
passed to eXist rather than the HTML that is returned, so the name of
the action has been changed to reflect this.  Additionally, this
implements an expanded set of rules for transforming a document that
should handle references to specific locations in other xml files.