Redefine .all_ography_types, add .all_ography_read_methods
It turns out that for validations it's useful to have a list of the
valid ography types as strings without '_for' appended to them.  Because
of this, I've changed .all_ography_types to return this list of strings
and added a new method called all_ography_read_methods that returns a
list of symbols ready to be sent to a CoreFile instance as messages.