Use #collections instead of #collection_ids
It turns out that #collection_ids is somewhat flaky, so to get around
potentially confusing failures this changes #all_ography_tei_files to
use the #collections.  This is noticeably slower but given that it is
only ever executed in the context of a background job this should be
acceptable for now.