Two quick fixes to make ographies work
1. The API now preprocesses params so that empty strings are deleted.
This means that it is now possible for our client to send us empty
(after preprocessing) arrays, which are used to implement some important
record/ography related functionality.

2. Important record/ography functionality that depends on empty arrays
no longer checks that the empty array is present (it's not) before doing
anything.  This means that it's now possible to set/unset a given record
as several ography types and then none.