Adds debugging output and fixes reading interface
There were two bugs with reading interface preparation that were causing
it to fail.  First, in the upsert_core_file action the Reading Interface
was being generated before page images were saved, which meant of course
that no references to these images could be found and that part of the
process would fail.  Second, it turns out that CoreFile instance we are
holding during prepare_reading_interface should be reloaded to ensure
that all relevant relationships that exist on it can be found when
queried.

Also, I left in the puts that I puts'd everywhere on the reading
interface code so that it'd be easy to see what was going on when
running Resque in a window on a dev environment.