Split works between maintenance and everything-else queues
Because we don't want tasks like rebuilding every reading interface in
the repository to lock up all of our workers, we put them on a separate
tapas_rails_maintenance queue.  This ensures that on deploy two workers
are listening on that queue and two are listening on the queue for
processing incoming requests from Drupal.