close pr #51, pr #27, pr #46, pr #45, pr #44, pr #42
This was weird, bc the PRs were submitted against develop, which was the
default branch. We needed to switch the default to rails_interface, and
so the merging got....weird. Instead of following the git signals, I
just went through the diffs on GitHub and manually did them. Hope I
didn't break things.