Adjust scroll button intersection observer: test the reader header.
Instead of testing the *page* header, we test the reader header for
an intersection threshold of 1.0. Thanks to a new `top: -1px` rule,
we now know that when the reader header is sticky, a part of it is
just out of the viewport.

I also set up an `.is-sticking` class on the reader header,
preparing to maybe reduce the amount of space the sticky header
takes up.