Wrap `.tapas-container` heading element in a <div>.
...and move the navigation links up next to that heading, using
Bootstrap's `.nav-tab` class for styling.

I did modify Bootstrap's styles so that the tab outline is always
visible, even if the link isn't active.

I also added border-bottom to links, rather than underlines, so
that we're not violating WCAG principles by relying only on color
to tell people that these are links.