Clean up routes a little.
Removed a couple routes I know we're getting rid of. Condensed the
routes for Projects, Collections, and Core Files. Also, added `:`
to identifiers in the route path definitions.

Eventually we'll need more complex URLs for projects, etc., but it
looks like Rails 7 has more powerful routing capabilities. It might
be good to wait until we've upgraded.