Move sort method checking into new method `set_up_browse`.
We'll need to check and sign off on the user's requested sort
method for the other Browse pages too, so the new method is in the
ListResources concern.

I also added a `sort_string` method for convenience in actually
performing a sort, and switched from camelCase to snake_case
variable names.