Rearrange top links and buttons on Browse page.
Added a summary of where we are in the list of paginated results,
and made the top pagination nav much simpler (just "Previous" and
"Next"). I swapped the upper pagination nav for the Expand/Collapse
controls, because visually, it makes more sense for the controls to
be near the "More information" toggles they'd be working on.