Reproduce the Browse page structure from the wireframe.
Since the heading is part of a tightly-styled container, I moved
its tag from the Project's template to the shared one. The
individual "Browse" pages will still set the content of their
headings, but will rely on the shared template for the HTML
encoding.

Pagination content and controls have been commented out or left
as empty elements, since I'll be implementing those in a separate
pass.