Randomize the number of 'ographies created; refine project roles.
Added a new method, `create_new_core_file()` to the dummy data
generator. It is called for a random number of core files (0 to 50)
and a random number of 'ography support files (0 to 2).

Also, `create_project_members()` is now closer to the project roles
described in issue #111. Only "owner" and "contributor" roles are
used, and only users who are project owners are made collection
depositors. The project depositor is not (yet?) made into an owner.