Set up variable to hold common sort options.
The ListResources concern (to be renamed when I can think of
something better) now contains a nested array of sort methods and
their human-readable labels. The concern also contains new methods
for setting and testing the sort parameters against the acceptable
values.

The Project controller adds the "number of TEI documents" sort to
the ones provided by the concern. (Implementation of the actual
sort provided by @candyhazlett!)

Finally, the shared browse page view uses the defined sort options
to populate the form field.